Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3905



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@16754 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/ucnvdisp.c
@@ -65,7 +65,7 @@ ucnv_getDisplayName(const UConverter *cnv,
     } else {
         /* convert the internal name into a Unicode string */
         *pErrorCode=U_ZERO_ERROR;
-        length=uprv_strlen(cnv->sharedData->staticData->name);
+        length=(int32_t)uprv_strlen(cnv->sharedData->staticData->name);
         u_charsToUChars(cnv->sharedData->staticData->name, displayName, uprv_min(length, displayNameCapacity));
     }
     return u_terminateUChars(displayName, displayNameCapacity, length, pErrorCode);
